Move all functions from main.7.c to main.c
